feat: added global design tokens based from Figma and SDD
