--- src/styles/tokens.css
@@ -1,47 +1,0 @@
-@theme {
-  /* Colors */
-  --color-primary: #C6623E;
-  --color-primary-hover: #A84E2E;
-  --color-primary-light: #F5E6E0;
-  --color-secondary: #95CDE0;
-  --color-accent: #8A6ED4;
-  --color-neutral: #FBF5F2;
-  --color-destructive: #DC2626;
-  --color-success: #16A34A;
-  --color-warning: #D97706;
-  --color-muted: #9CA3AF;
-  --color-surface-light: #FFFFFF;
-  --color-surface-dark: #1A1A1A;
-
-  /* Typography */
-  --font-heading: 'Plus Jakarta Sans', 'Space Grotesk', sans-serif;
-  --font-body: 'Inter', sans-serif;
-
-  /* Font sizes */
-  --text-h1: 45px;
-  --text-h2: 35px;
-  --text-h3: 25px;
-  --text-h4: 18px;
-  --text-h5: 16px;
-  --text-paragraph: 16px;
-
-  /* Spacing (4px base) */
-  --spacing-1: 4px;
-  --spacing-2: 8px;
-  --spacing-3: 12px;
-  --spacing-4: 16px;
-  --spacing-6: 24px;
-  --spacing-8: 32px;
-  --spacing-10: 40px;
-  --spacing-12: 48px;
-  --spacing-16: 64px;
-
-  /* Base tokens */
-  --radius: 8px;
-  --radius-full: 9999px;
-  --sidebar-width: 260px;
-  --card-padding: 24px;
-  --grid-max-width: 1440px;
-  --grid-gutter: 24px;
-  --grid-margin: 80px;
-}
